@@ -69,7 +69,7 @@ export class SDKBrowserModule {
   static forRoot(internalStorageProvider: any = {
     provide: InternalStorage,
     useClass: CookieBrowser
-  }): ModuleWithProviders {
+  }): ModuleWithProviders<SDKBrowserModule> {
     return {
       ngModule  : SDKBrowserModule,
       providers : [
